+++ docs/include/samplecode/SpotifyClient/CheckTrackFavorites.py
@@ -25,6 +25,14 @@ try:
     for key in result.keys():
         print('- %s = %s' % (key, result[key]))
 
+    # check if nowplaying track is already saved in the current Spotify user's 'Your Library'.
+    print('\nChecking if nowplaying track is saved by the current user ...')
+    result:dict = spotify.CheckTrackFavorites()
+            
+    print('\nResults:')
+    for key in result.keys():
+        print('- %s = %s' % (key, result[key]))
+
 except Exception as ex:
 
     print("** Exception: %s" % str(ex))